Proyecto avance

+++ ProyectoRE/src/main/java/com/ProyectoRE/controller/CitaController.java
@@ -1,13 +1,17 @@
 package com.ProyectoRE.controller;
 
 import com.ProyectoRE.domain.Cita;
+import com.ProyectoRE.domain.Usuario;
+import com.ProyectoRE.domain.Propiedad;
 import com.ProyectoRE.service.CitaService;
+import com.ProyectoRE.service.UsuarioService;
+import com.ProyectoRE.service.PropiedadService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @Controller
 @RequestMapping("/cita")
@@ -16,30 +20,41 @@ public class CitaController {
     @Autowired
     private CitaService citaService;
 
+    @Autowired
+    private UsuarioService usuarioService;
+
+    @Autowired
+    private PropiedadService propiedadService;
+
     @GetMapping("/formulario")
-    public String mostrarFormularioCitas() {
-        return "citas-form"; // Renderiza citas-form.html
+    public String mostrarFormularioCita(Model model) {
+        List<Usuario> usuarios = usuarioService.getUsuarios();
+        List<Propiedad> propiedades = propiedadService.getPropiedades(true); // Obtener solo propiedades activas
+        model.addAttribute("usuarios", usuarios);
+        model.addAttribute("propiedades", propiedades);
+        model.addAttribute("cita", new Cita()); // Crear un objeto vacío para el formulario
+        return "cita/formulario";
     }
 
     @PostMapping("/guardar")
-    public String guardarCita(Cita cita) {
+    public String guardarCita(@ModelAttribute Cita cita) {
         citaService.save(cita);
         return "redirect:/cita/listado";
     }
 
     @GetMapping("/listado")
-    public String listadoCitas(Model model) {
-        var citas = citaService.getCitas();
+    public String listarCitas(Model model) {
+        List<Cita> citas = citaService.getCitas();
         model.addAttribute("citas", citas);
-        model.addAttribute("totalCitas", citas.size());
-        return "citas"; // Renderiza citas.html
+        return "cita/listado";
     }
 
-    @GetMapping("/eliminar/{idCita}")
-    public String eliminarCita(Cita cita) {
-        citaService.delete(cita);
+    @GetMapping("/eliminar/{id}")
+    public String eliminarCita(@PathVariable("id") int id) {
+        var cita = citaService.findById(id);
+        if (cita != null) {
+            citaService.delete(cita);
+        }
         return "redirect:/cita/listado";
     }
 }
-
-
